feat: refactorizar la funcionalidad de búsqueda de productos e integrarla en los componentes de compra
