+++ client/src/app/_components/ventas/punto-venta.component.html
@@ -8,21 +8,14 @@
     </div>
 
     <div class="bg-white rounded-xl shadow-lg border border-amber-200 p-6 mb-6">
-      <div class="flex flex-col md:flex-row md:items-end gap-4">
-        <div class="flex-grow">
-          <label for="sku" class="block text-sm font-medium text-amber-700 mb-1">
-            Código de Barras / SKU
-          </label>
-          <input type="text" id="sku" name="sku" #skuInputRef [(ngModel)]="skuInput" (keyup.enter)="onSkuSubmit()"
-            class="w-full px-4 py-2 border border-amber-300 rounded-lg shadow-sm focus:outline-none focus:ring-2 focus:ring-amber-500 focus:border-transparent"
-            placeholder="Escanea o ingresa el código" [disabled]="isLoading" autocomplete="off">
-        </div>
-        <button (click)="onSkuSubmit()"
-          class="px-6 py-2 bg-amber-600 hover:bg-amber-700 text-white font-medium rounded-lg shadow-sm focus:outline-none focus:ring-2 focus:ring-amber-500 focus:ring-offset-2"
-          [disabled]="isLoading">
-          {{ isLoading ? 'Buscando...' : 'Buscar' }}
-        </button>
-      </div>
+      <app-product-search
+        (productSelected)="onProductSelected($event)"
+        placeholder="Escanea o ingresa el código, nombre o SKU"
+        label="Buscar Producto (Código de Barras / SKU / Nombre)"
+        [showPriceInfo]="true"
+        [showStockInfo]="true"
+        [showDescription]="true">
+      </app-product-search>
     </div>
     <div class="bg-white rounded-xl shadow-lg border border-amber-200 p-6 mb-6">
       <h2 class="text-xl font-bold text-amber-900 mb-4">Productos en la venta</h2>
